+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [`PowAssent.Phoenix.AuthorizationController`] Now logs error for when user can't be created
 
+### Bug fixes
+
+* [`PowAssent.Phoenix.AuthorizationController`] Fixed bug where registration path couldn't be found due to no route helpers
+
 ## v0.4.16 (2023-03-21)
 
 Now uses Phoenix 1.7 components templating and requires Pow 1.0.29.